Fixed cast warnings in c++
git-svn-id: svn://136.177.114.72/svn_GW/phreeqcpp/trunk/src@1008 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ Dictionary.cxx
@@ -56,7 +56,7 @@ int cxxDictionary::putString(std::string str) {
 	if (it != stringkey.end()) {
 		n = it->second;
 	} else {
-		n = stringkey.size();
+		n = (int) stringkey.size();
 		stringkey[str] = n;
 		intkey[n] = str;
 	}
